+++ vitest.workspace.ts
@@ -1,3 +1,7 @@
 import { defineWorkspace } from "vitest/config";
 
-export default defineWorkspace(["packages/domain", "packages/application"]);
+export default defineWorkspace([
+  "packages/domain",
+  "packages/application",
+  "packages/infrastructure",
+]);